cleanup: rename socket filter prog in netolly
If you run together netolly and appolly you will see 2 ebpf program with
the same name `beyla_socket__http_filter` this is not ideal if you want
to debug/benchamrk ebpf progs. Moreover this is not an http specific
code so it could make sense to change its name in any case.

Signed-off-by: Andrea Terzolo <andreaterzolo3@gmail.com>